requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
av==15.1.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.2.2
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.12.0
livekit==1.0.18
livekit-agents==1.2.18
livekit-api==1.0.7
livekit-blingfire==1.0.0
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-silero==1.2.18
livekit-plugins-turn-detector==1.2.18
livekit-protocol==1.1.0
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.3.4
onnxruntime==1.23.1
openai==2.7.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
prometheus_client==0.23.1
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
sniffio==1.3.1
sounddevice==0.5.3
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
types-protobuf==6.32.1.20251105
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
watchfiles==1.1.1
yarl==1.22.0
zipp==3.23.0
